proj22/requirements.txt
absl-py==0.11.0
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
base58==2.0.1
bleach==3.2.1
blinker==1.4
boto3==1.16.10
botocore==1.19.10
cachetools==4.1.1
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1598400192773/work
colorama==0.4.4
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
enum-compat==0.0.3
fsspec==0.8.4
future @ file:///D:/bld/future_1605203720223/work
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
gym @ file:///D:/bld/gym_1607105517412/work
idna==2.10
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kornia==0.4.1
Markdown==3.3.3
MarkupSafe==1.1.1
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
notebook==6.1.4
numpy @ file:///C:/ci/numpy_and_numpy_base_1603466732592/work
oauthlib==3.1.0
olefile==0.46
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pathtools==0.1.2
pickleshare==0.7.5
Pillow @ file:///C:/ci/pillow_1603823068645/work
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.5.0
pyglet==1.5.0
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-lightning==1.0.5
pytz==2020.4
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
scipy @ file:///C:/bld/scipy_1604304921792/work
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.4
streamlit==0.70.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
terminado==0.9.1
testpath==0.4.4
toml==0.10.2
toolz==0.11.1
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
tzlocal==2.1
urllib3==1.25.11
validators==0.18.1
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
